+++ packages/vscode-tailwindcss/src/extension.ts
@@ -10,26 +10,19 @@ import type {
   TextEditorDecorationType,
   ConfigurationScope,
   WorkspaceConfiguration,
-  CompletionList,
-  ProviderResult,
   Selection,
 } from 'vscode'
 import {
   workspace as Workspace,
   window as Window,
-  languages as Languages,
   Uri,
   commands,
   SymbolInformation,
   Position,
   Range,
   RelativePattern,
-  CompletionItem,
-  CompletionItemKind,
-  SnippetString,
-  TextEdit,
 } from 'vscode'
-import type { LanguageClientOptions, ServerOptions, Disposable } from 'vscode-languageclient/node'
+import type { LanguageClientOptions, ServerOptions } from 'vscode-languageclient/node'
 import {
   LanguageClient,
   TransportKind,
@@ -45,6 +38,7 @@ import picomatch from 'picomatch'
 import { CONFIG_GLOB, CSS_GLOB } from '@tailwindcss/language-server/src/lib/constants'
 import braces from 'braces'
 import normalizePath from 'normalize-path'
+import * as servers from './servers/index'
 
 const colorNames = Object.keys(namedColors)
 
@@ -348,114 +342,6 @@ export async function activate(context: ExtensionContext) {
     }),
   )
 
-  let cssServerBooted = false
-  async function bootCssServer() {
-    if (cssServerBooted) return
-    cssServerBooted = true
-
-    let module = context.asAbsolutePath(path.join('dist', 'cssServer.js'))
-    let prod = path.join('dist', 'tailwindModeServer.js')
-
-    try {
-      await Workspace.fs.stat(Uri.joinPath(context.extensionUri, prod))
-      module = context.asAbsolutePath(prod)
-    } catch (_) {}
-
-    let client = new LanguageClient(
-      'tailwindcss-intellisense-css',
-      'Tailwind CSS',
-      {
-        run: {
-          module,
-          transport: TransportKind.ipc,
-        },
-        debug: {
-          module,
-          transport: TransportKind.ipc,
-          options: {
-            execArgv: ['--nolazy', '--inspect=6051'],
-          },
-        },
-      },
-      {
-        documentSelector: [{ language: 'tailwindcss' }],
-        outputChannelName: 'Tailwind CSS Language Mode',
-        synchronize: { configurationSection: ['css'] },
-        middleware: {
-          provideCompletionItem(document, position, context, token, next) {
-            function updateRanges(item: CompletionItem) {
-              const range = item.range
-              if (
-                range instanceof Range &&
-                range.end.isAfter(position) &&
-                range.start.isBeforeOrEqual(position)
-              ) {
-                item.range = { inserting: new Range(range.start, position), replacing: range }
-              }
-            }
-            function updateLabel(item: CompletionItem) {
-              if (item.kind === CompletionItemKind.Color) {
-                item.label = {
-                  label: item.label as string,
-                  description: item.documentation as string,
-                }
-              }
-            }
-            function updateProposals(
-              r: CompletionItem[] | CompletionList | null | undefined,
-            ): CompletionItem[] | CompletionList | null | undefined {
-              if (r) {
-                ;(Array.isArray(r) ? r : r.items).forEach(updateRanges)
-                ;(Array.isArray(r) ? r : r.items).forEach(updateLabel)
-              }
-              return r
-            }
-            const isThenable = <T>(obj: ProviderResult<T>): obj is Thenable<T> =>
-              obj && (<any>obj)['then']
-
-            const r = next(document, position, context, token)
-            if (isThenable<CompletionItem[] | CompletionList | null | undefined>(r)) {
-              return r.then(updateProposals)
-            }
-            return updateProposals(r)
-          },
-        },
-      },
-    )
-
-    await client.start()
-    context.subscriptions.push(initCompletionProvider())
-
-    function initCompletionProvider(): Disposable {
-      const regionCompletionRegExpr = /^(\s*)(\/(\*\s*(#\w*)?)?)?$/
-
-      return Languages.registerCompletionItemProvider(['tailwindcss'], {
-        provideCompletionItems(doc: TextDocument, pos: Position) {
-          let lineUntilPos = doc.getText(new Range(new Position(pos.line, 0), pos))
-          let match = lineUntilPos.match(regionCompletionRegExpr)
-          if (match) {
-            let range = new Range(new Position(pos.line, match[1].length), pos)
-            let beginProposal = new CompletionItem('#region', CompletionItemKind.Snippet)
-            beginProposal.range = range
-            TextEdit.replace(range, '/* #region */')
-            beginProposal.insertText = new SnippetString('/* #region $1*/')
-            beginProposal.documentation = 'Folding Region Start'
-            beginProposal.filterText = match[2]
-            beginProposal.sortText = 'za'
-            let endProposal = new CompletionItem('#endregion', CompletionItemKind.Snippet)
-            endProposal.range = range
-            endProposal.insertText = '/* #endregion */'
-            endProposal.documentation = 'Folding Region End'
-            endProposal.sortText = 'zb'
-            endProposal.filterText = match[2]
-            return [beginProposal, endProposal]
-          }
-          return null
-        },
-      })
-    }
-  }
-
   function bootWorkspaceClient(folder: WorkspaceFolder) {
     if (clients.has(folder.uri.toString())) {
       return
@@ -511,7 +397,7 @@ export async function activate(context: ExtensionContext) {
     }
 
     let clientOptions: LanguageClientOptions = {
-      documentSelector: languages.get(folder.uri.toString()).map((language) => ({
+      documentSelector: languages.get(folder.uri.toString())!.map((language) => ({
         scheme: 'file',
         language,
         pattern: normalizePath(`${folder.uri.fsPath.replace(/[\[\]\{\}]/g, '?')}/**/*`),
@@ -770,7 +656,7 @@ export async function activate(context: ExtensionContext) {
 
   async function didOpenTextDocument(document: TextDocument): Promise<void> {
     if (document.languageId === 'tailwindcss') {
-      bootCssServer()
+      servers.css.boot(context, outputChannel)
     }
 
     // We are only interested in language mode text
